changing pattern to match section 1
